changes Seedshop filename
