--- Seedshop/SeedshopRevised.tscn
@@ -1,98 +1,0 @@
-[gd_scene load_steps=3 format=3 uid="uid://bd54v5orqpb6t"]
-
-[ext_resource type="Script" path="res://Seedshop/SeedshopRevised.cs" id="1_frgxh"]
-[ext_resource type="PackedScene" uid="uid://w84710wgdj5r" path="res://Seedshop/ShopSlot.tscn" id="1_hct0p"]
-
-[node name="SeedshopRevised" type="CanvasLayer"]
-script = ExtResource("1_frgxh")
-
-[node name="ColorRect" type="ColorRect" parent="."]
-offset_right = 502.0
-offset_bottom = 702.0
-color = Color(0.39, 0.39, 0.39, 1)
-
-[node name="GridContainer" type="GridContainer" parent="ColorRect"]
-layout_mode = 1
-anchors_preset = 8
-anchor_left = 0.5
-anchor_top = 0.5
-anchor_right = 0.5
-anchor_bottom = 0.5
-offset_left = -226.0
-offset_top = -326.0
-offset_right = 226.0
-offset_bottom = 326.0
-grow_horizontal = 2
-grow_vertical = 2
-columns = 4
-
-[node name="Panel1" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel2" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel3" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel4" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel5" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel6" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel7" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel8" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel9" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel10" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel11" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel12" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel13" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel14" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel15" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="Panel16" parent="ColorRect/GridContainer" instance=ExtResource("1_hct0p")]
-custom_minimum_size = Vector2(110, 160)
-layout_mode = 2
-
-[node name="CloseButton" type="Button" parent="."]
-offset_left = 548.0
-offset_top = 18.0
-offset_right = 600.0
-offset_bottom = 70.0
-text = "Close"
