--- TestScenes/TestSeedshopScene.tscn
@@ -1,21 +1,0 @@
-[gd_scene load_steps=5 format=3 uid="uid://blnwnotey1wvy"]
-
-[ext_resource type="PackedScene" uid="uid://duxkc2whc5lx3" path="res://Player/Player.tscn" id="1_rnhb8"]
-[ext_resource type="PackedScene" uid="uid://6m0jv0pp1br6" path="res://NPC/NPC.tscn" id="2_l6ykw"]
-[ext_resource type="Script" path="res://NPC/Seedboy.cs" id="3_h5vj8"]
-[ext_resource type="PackedScene" uid="uid://cg5yho3bnwg55" path="res://Seedshop/Seedshop.tscn" id="4_povm7"]
-
-[node name="TestInventoryScene" type="Node2D"]
-
-[node name="Player" parent="." instance=ExtResource("1_rnhb8")]
-position = Vector2(28, 26)
-
-[node name="Seedboy" parent="." instance=ExtResource("2_l6ykw")]
-position = Vector2(55, 113)
-scale = Vector2(1.00188, 0.98962)
-script = ExtResource("3_h5vj8")
-ActionName = "shop"
-
-[node name="Control" parent="." instance=ExtResource("4_povm7")]
-scale = Vector2(0.995229, 1.00289)
-transform = Transform2D(0.995229, 0, 0, 1.00289, 0, 0)
